+++ helper/local_context.go
@@ -67,6 +67,10 @@ func (p *LocalContext) GetRuntimeContext() context.Context {
 	return p.ctx
 }
 
+func (p *LocalContext) GetExtension(name string, cfg any) (pipeline.Extension, error) {
+	return nil, nil
+}
+
 func (p *LocalContext) RegisterCounterMetric(metric pipeline.CounterMetric) {
 	contextMutex.Lock()
 	defer contextMutex.Unlock()