+++ pkg/pipeline/extensions/decoder.go
@@ -1,0 +1,32 @@
+// Copyright 2021 iLogtail Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package extensions
+
+import (
+	"net/http"
+
+	"github.com/alibaba/ilogtail/pkg/models"
+	"github.com/alibaba/ilogtail/pkg/protocol"
+)
+
+// Decoder used to parse buffer to sls logs
+type Decoder interface {
+	// Decode reader to logs
+	Decode(data []byte, req *http.Request, tags map[string]string) (logs []*protocol.Log, err error)
+	// DecodeV2 reader to groupEvents
+	DecodeV2(data []byte, req *http.Request) (groups []*models.PipelineGroupEvents, err error)
+	// ParseRequest gets the request's body raw data and status code.
+	ParseRequest(res http.ResponseWriter, req *http.Request, maxBodySize int64) (data []byte, statusCode int, err error)
+}